# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CelebFanCards API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Celebrity {
  id: number;
  name: string;
  bio: string;
  category: string;
  imageUrl: string;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  popularFor?: string | null;
  fanCount: number;
  createdAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface Order {
  id: number;
  confirmationNumber: string;
  celebId: number;
  celebName: string;
  cardType: string;
  price: number;
  status: string;
  fullName: string;
  email: string;
  username: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  specialMessage?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  transactionId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateOrderBodyCardType =
  (typeof CreateOrderBodyCardType)[keyof typeof CreateOrderBodyCardType];

export const CreateOrderBodyCardType = {
  basic: "basic",
  premium: "premium",
  vip: "vip",
} as const;

export interface CreateOrderBody {
  celebId: number;
  cardType: CreateOrderBodyCardType;
  fullName: string;
  email: string;
  username: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  specialMessage?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  processing: "processing",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  transactionId?: string | null;
}

export interface User {
  id: number;
  fullName: string;
  email: string;
  username: string;
  createdAt: string;
}

export interface SignUpBody {
  fullName: string;
  email: string;
  username: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface StatsSummary {
  totalCelebrities: number;
  totalOrders: number;
  totalFans: number;
  totalRevenue: number;
}

export interface PopularCelebrity {
  id: number;
  name: string;
  imageUrl: string;
  category: string;
  orderCount: number;
}

export interface InitializePaymentBody {
  orderId: number;
  amount: number;
  currency: string;
  email: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  redirectUrl: string;
}

export interface PaymentInitResponse {
  paymentLink: string;
  txRef: string;
}

export interface VerifyPaymentBody {
  transactionId: string;
  orderId: number;
}

export interface PaymentVerifyResponse {
  status: string;
  message: string;
  orderId: number;
  confirmationNumber: string;
}

export type ListCelebritiesParams = {
  category?: string;
  search?: string;
};

export type ListOrdersParams = {
  userEmail?: string;
};