# lib/api-zod/src/generated/types/celebrity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CelebFanCards API
 * OpenAPI spec version: 0.1.0
 */

export interface Celebrity {
  id: number;
  name: string;
  bio: string;
  category: string;
  imageUrl: string;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  popularFor?: string | null;
  fanCount: number;
  createdAt: string;
}