# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CelebFanCards API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyCardType } from "./createOrderBodyCardType";

export interface CreateOrderBody {
  celebId: number;
  cardType: CreateOrderBodyCardType;
  fullName: string;
  email: string;
  username: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  specialMessage?: string | null;
}