# lib/api-zod/src/generated/types/paymentInitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CelebFanCards API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInitResponse {
  paymentLink: string;
  txRef: string;
}