# lib/api-zod/src/generated/types/signUpBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CelebFanCards API
 * OpenAPI spec version: 0.1.0
 */

export interface SignUpBody {
  fullName: string;
  email: string;
  username: string;
  password: string;
}